Renames 'db' to 'util' and updates references; fixes bug in trip_detail
Moves all functions from 'app', which are not route related, to
'util'. This makes the two files more distinct in their purpose and
makes it easier to find what you need, dependent on the nature of
what you're looking for.

Updates 'forms' to import from 'util' instead of 'db'.

The trip_detailed view within 'app' was not retaining the End Date
from the previous stop; added variable to fix this.
